pyproject.toml
[project]
name = "mesa-certa"
version = "0.1.0"
description = "Rede social de avaliação de restaurantes"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-migrate>=4.0.0",
    "flask-login>=0.6.0",
    "flask-wtf>=1.3.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.1.0",
    "email-validator>=2.0.0",
    "flask-limiter>=4.1.1",
    "flask-talisman>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "pytest>=9.0.0",
    "pytest-flask>=1.3.0",
    "pytest-cov>=7.1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"seed.py" = ["E501"]
"tests/**" = ["E501"]
